Refactoring all the store files
