+++ src/main/resources/watguessr/src/stores/index.ts
@@ -1,13 +1,12 @@
 import { createStore } from 'vuex';
 import type { Module } from 'vuex';
-import { gameModule } from './modules/game';
-import roundModule from './modules/round';
-import guessModule from './modules/guess';
+import { gameModule } from './modules/game/index';
+import roundModule from './modules/round/index';
+import guessModule from './modules/guess/index';
 import leaderboardModule from './modules/leaderboard';
-import type { RoundState } from './modules/round';
-import type { GameState } from './modules/game';
-import type { GuessState } from './modules/guess';
-import sceneModule, {type SceneState} from "@/stores/modules/scene.ts";
+import type { RoundState } from './modules/round/state';
+import type { GameState } from './modules/game/state';
+import type { GuessState } from './modules/guess/state';
 
 export interface LeaderboardState {
   leaderboardData: any;
@@ -21,7 +20,6 @@ export interface RootState {
   game: GameState;
   guess: GuessState;
   leaderboard: LeaderboardState;
-  scene: SceneState;
 }
 
 const store = createStore<RootState>({
@@ -30,7 +28,6 @@ const store = createStore<RootState>({
     round: roundModule as Module<RoundState, RootState>,
     guess: guessModule as Module<GuessState, RootState>,
     leaderboard: leaderboardModule as Module<LeaderboardState, RootState>,
-    scene: sceneModule as Module<SceneState, RootState>
   },
 });
