--- src/main/resources/watguessr/src/stores/modules/game.ts
@@ -1,142 +1,0 @@
-// src/stores/modules/game.ts
-import type { Module, ActionTree, MutationTree, GetterTree } from 'vuex';
-import type { RootState } from '../index';
-
-// Define GameState
-export interface GameState {
-  gameId: string | null;
-  gameMode: string;
-  status: 'idle' | 'loading' | 'playing' | 'ended';
-  currentRound: number;
-  maxRounds: number;
-  scores: Record<string, number>;
-  winner: string | null;
-  currentView: string
-}
-
-// Initial State
-const state: GameState = {
-  gameId: null,
-  gameMode: '',
-  status: 'idle',
-  currentRound: 1,
-  maxRounds: 5,
-  scores: {},
-  winner: null,
-  currentView: 'Map'
-};
-
-// Define mutations with proper typing
-const mutations: MutationTree<GameState> = {
-  SET_GAME_ID(state, gameId: string) {
-    state.gameId = gameId;
-  },
-  SET_GAME_MODE(state, mode: string) {
-    state.gameMode = mode;
-  },
-  SET_STATUS(state, status: GameState['status']) {
-    state.status = status;
-  },
-  INCREMENT_ROUND(state) {
-    state.currentRound++;
-  },
-  CHANGE_VIEW(state, nextView: string) {
-    state.currentView = nextView;
-  },
-  ADD_SCORE(state, payload: {username: string, score: number}) {
-    if (!state.scores[payload.username]) {
-      state.scores[payload.username] = 0;
-    }
-    state.scores[payload.username] += payload.score;
-  },
-  SET_FINAL_WINNER(state, winner: string) {
-    state.winner = winner;
-    state.status = 'ended';
-  },
-  RESET_GAME(state) {
-    state.gameId = null;
-    state.status = 'idle';
-    state.currentRound = 1;
-    state.scores = {};
-    state.winner = null;
-  },
-};
-
-// Define actions with proper typing
-const actions: ActionTree<GameState, RootState> = {
-  async createSingleplayerGame({ commit, dispatch }) {
-    commit('RESET_GAME');
-    commit('SET_STATUS', 'loading');
-    commit('SET_GAME_MODE', 'Singleplayer');
-
-    const response = await fetch('http://localhost:5173/api/game/create/singleplayer');
-    const gameId = await response.json();
-
-    commit('SET_GAME_ID', gameId);
-    dispatch('round/startRound', {gameId: gameId}, {root: true});
-    commit('SET_STATUS', 'playing');
-  },
-
-  recordRoundWinner({ state, commit, dispatch }, payload: {username: string, score: number}) {
-    commit('ADD_SCORE', payload);
-    commit('CHANGE_VIEW', 'RoundEnd');
-
-    if (state.currentRound >= state.maxRounds) {
-      const winner = Object.entries(state.scores).sort((a, b) => b[1] - a[1])[0][0];
-      //TODO: send api call to set final winner
-
-      commit('SET_FINAL_WINNER', winner);
-    } else {
-      dispatch('round/resetRound', null, { root: true });
-
-      const newScene = generateNextScene(); // Replace with actual scene logic
-      dispatch('round/startRound', newScene, { root: true });
-      commit('INCREMENT_ROUND');
-    }
-  },
-
-  async endGame({state, commit}, gameId: string) {
-    commit('RESET_GAME');
-
-    try {
-      const response = await fetch(`http://localhost:5173/api/game/finish/singleplayer?gameId=${gameId}`, {
-        method: 'POST'
-      })
-
-      if (!response.ok) {
-        throw new Error(`HTTP error! status: ${response.status}`)
-      }
-
-      const result: number = await response.json()
-      return result
-    } catch (error) {
-      console.error('Error finishing game:', error)
-      throw error
-    }
-  }
-};
-
-// Define getters with proper typing
-const getters: GetterTree<GameState, RootState> = {
-  gameId: (state) => state.gameId,
-  gameMode: (state) => state.gameMode,
-  currentRound: (state) => state.currentRound,
-  gameStatus: (state) => state.status,
-  finalWinner: (state) => state.winner,
-  scores: (state) => state.scores,
-  currentView: (state) => state.currentView
-};
-
-export const gameModule: Module<GameState, RootState> = {
-  namespaced: true,
-  state,
-  mutations,
-  actions,
-  getters,
-};
-
-// Dummy placeholder for now
-function generateNextScene() {
-  //TODO: generating next scene logic
-  return {}; // Replace with actual scene generation logic
-}
