+++ src/main/resources/watguessr/src/stores/modules/game/getters.ts
@@ -1,0 +1,14 @@
+// src/stores/modules/game/getters.ts
+import type { GetterTree } from 'vuex';
+import type { RootState } from '../../index';
+import type { GameState } from './state';
+
+export const getters: GetterTree<GameState, RootState> = {
+  getGameId: (state) => state.gameId,
+  getGameMode: (state) => state.gameMode,
+  getCurrentRound: (state) => state.currentRound,
+  getGameStatus: (state) => state.status,
+  getFinalWinner: (state) => state.winner,
+  getScores: (state) => state.scores,
+  getCurrentView: (state) => state.currentView
+};
