+++ src/main/resources/watguessr/src/stores/modules/guess/state.ts
@@ -1,0 +1,6 @@
+// src/stores/modules/guess/state.ts
+export interface GuessState {
+  // Currently no local state
+}
+
+export const state = (): GuessState => ({});
